Refactor curated_variables.py to cast columns to correct data types

+++ big_data/workflows/spark/curated/curated_variables.py
@@ -9,10 +9,10 @@ def transform(ec):
     df_variables = ec.config["data"]["dbw_variables_extract"]
     
     df_transform = (df_variables
-                    .withColumn("id_dim_1", F.concat(F.col("id_wymiar_1"), F.col("id_pozycja_1")))
-                    .withColumn("id_dim_2", F.concat(F.col("id_wymiar_2"), F.col("id_pozycja_2")))
-                    .withColumn("id_dim_3", F.concat(F.col("id_wymiar_3"), F.col("id_pozycja_3")))
-                    .withColumn("id_dim_4", F.concat(F.col("id_wymiar_4"), F.col("id_pozycja_4")))
+                    .withColumn("id_dim_1", F.concat(F.col("id_wymiar_1"), F.col("id_pozycja_1")).cast("bigint"))
+                    .withColumn("id_dim_2", F.concat(F.col("id_wymiar_2"), F.col("id_pozycja_2")).cast("bigint"))
+                    .withColumn("id_dim_3", F.concat(F.col("id_wymiar_3"), F.col("id_pozycja_3")).cast("bigint"))
+                    .withColumn("id_dim_4", F.concat(F.col("id_wymiar_4"), F.col("id_pozycja_4")).cast("bigint"))
                     .select(
                         F.col("id_zmienna").alias("id_zmienna"),
                         F.col("id_przekroj").alias("id_przekroj"),
